Separate flux computation to separate file
