+++ applications/DG-Max/CMakeLists.txt
@@ -32,6 +32,7 @@ if(hpGEM_USE_COMPLEX_PETSC AND hpGEM_USE_SLEPC)
 		Utils/KPhaseShift.cpp
 		Utils/CGDGMatrixKPhaseShiftBuilder.cpp
 		Utils/FaceKPhaseShiftBuilder.cpp
+		Utils/FluxFacets.cpp
 		Utils/PredefinedStructure.cpp
 		Utils/StructureDescription.cpp
 		Utils/ZoneStructureDescription.cpp
